fix(test): restore domain verification entrypoints
Restore the approved resolver, privacy, and lifecycle unittest paths while keeping split scenarios discoverable exactly once. Remove the test-only production JSON reader and close project-store descriptors through a finally-owned construction boundary.

Constraint: Preserve exact Todo 5, Todo 8, Todo 10, and F3 verification commands without duplicate discovery runs.
Rejected: Re-export TestCase subclasses from every split module | unittest discovery would execute the same scenarios more than once.
Confidence: high
Scope-risk: narrow
Directive: Keep compatibility mixins outside TestCase inheritance and preserve the lazy resolver compatibility export.
Tested: exact Todo 5/8/10/F3 commands; 2811-test discovery; broad-exception policy; compileall; docs check; Ruff; AST pure LOC; diff check.
Not-tested: Remote CI was not run.
